Union for mat4 struct added

+++ Spraky-core/x64/src/maths/maths_func.h
@@ -1,12 +1,12 @@
 #pragma once
-	
+
 #define _USE_MATH_DEFINES
 #include <math.h>
 
 namespace sparky {
 	namespace maths {
 		inline float toRadians(float degrees) {
-			return degrees * M_PI / 180.0f;
+			return static_cast<float>(degrees * M_PI / 180.0f);
 		}
 	}
 }
